# src/main/resources/static/api/v1/openapi3.yaml
openapi: 3.0.1
info:
  title: Exam Lab API
  description: API for Exam Lab
  version: 0.1.0
servers:
- url: https://exam-lab.store/
- url: http://localhost:8080/
...
paths:
...
  /api/v1/questions:
    put:
      tags:
      - question
      summary: Update question
      description: Question 내용 업데이트
      operationId: update-question
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-questions-420259584'
            examples:
              update-question:
                value: "{\"question\":\"변경된 질문입니다.\",\"options\":[\"변경된 보기1\",\"변경\
                  된 보기2\",\"변경된 보기3\",\"변경된 보기4\"],\"answers\":[1,3],\"id\":\"98a6a8b3-3b64-4d53-999a-003b86aeb0af\"\
                  ,\"commentary\":\"변경된 설명입니다.\",\"tags\":{\"난이도\":[\"중\"],\"단원\"\
                  :[\"2\"]}}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-questions486549215'
              examples:
                update-question:
                  value: "{\"code\":200,\"message\":\"OK\"}"
...
  /api/v1/questions/{questionId}:
    delete:
      tags:
      - question
      summary: Delete questions by questionId
      description: 해당되는 문제 ID Question 삭제
      operationId: delete-questions-by-questionId
      parameters:
      - name: questionId
        in: path
        description: QuestionId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-questions486549215'
              examples:
                delete-questions-by-questionId:
                  value: "{\"code\":200,\"message\":\"OK\"}"
...
  /api/v1/exams/{examId}/questions:
    get:
      tags:
      - question
      summary: Search questions
      description: Exam ID를 통한 Questions 조회
      operationId: Search-questions
      parameters:
      - name: examId
        in: path
        description: ExamId
        required: true
        schema:
          type: integer
          format: int32
      - name: tags_단원
        in: query
        description: 문제의 태그 - 단원
        required: false
        schema:
          type: string
      - name: tags_난이도
        in: query
        description: 문제의 태그 - 난이도
        required: false
        schema:
          type: string
      - name: includes
        in: query
        description: 문제에 포함된 키워드
        required: false
        schema:
          type: string
      - name: count
        in: query
        description: 검색된 문제의 개수
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-exams-examId-questions-1876066633'
              examples:
                Search-questions:
                  value: "{\"questions\":[{\"id\":\"3aab0f4a-5d26-4bca-b680-2ad3ec1cdbb6\"\
                    ,\"type\":\"객관식\",\"question\":\"소웨공 테스트 문제?\",\"question_images_in\"\
                    :[],\"question_images_out\":[],\"options\":[\"① 답 예시 1\",\"② 답\
                    \ 예시 2\",\"③ 답 예시 3\",\"④ 답 예시 4\"],\"answers\":[3],\"commentary\"\
                    :\"30이 가장 큰 수입니다.\",\"commentary_images_in\":[],\"commentary_images_out\"\
                    :[],\"tags\":{\"난이도\":[\"하\"],\"단원\":[\"1\"]}}],\"size\":1}"
    post:
      tags:
      - question
      summary: Add question
      description: 문제 추가
      operationId: add-questions
      parameters:
      - name: examId
        in: path
        description: ""
        required: true
        schema:
          type: string
      - name: Content-Type
        in: header
        description: |-
          questionUploadInfo-ContentType: application/jsonquestionImagesIn-ContentType:image.png
          questionImagesOut-ContentType:image.png
          commentaryImagesOut-ContentType:image.png
          commentaryImagesOut-ContentType:image.png
          multipart/form-data
        required: true
        schema:
          type: string
        example: multipart/form-data
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-questions486549215'
              examples:
                add-questions:
                  value: "{\"code\":201,\"message\":{\"id\":\"41706e43-60fb-4a9e-9781-dba92c56d145\"\
                    ,\"type\":\"객관식\",\"question\":\"다음 중 총중량 1.5톤 피견인 승용자동차를 4.5톤\
                    \ 화물자동차로 견인하는 경우 필요한 운전면허에 해당하지 않은 것은?\",\"question_images_in\"\
                    :[{\"url\":\"https://examlab-image.s3.ap-northeast-2.amazonaws.com/test_images/image1.png\"\
                    ,\"description\":\"설명1\",\"attribute\":\"속성1\"}],\"question_images_out\"\
                    :[{\"url\":\"https://examlab-image.s3.ap-northeast-2.amazonaws.com/test_images/image2.png\"\
                    ,\"description\":\"설명2\",\"attribute\":\"속성2\"}],\"options\":[\"\
                    ① 제1종 대형면허 및 소형견인차면허\",\"② 제1종 보통면허 및 대형견인차면허\",\"③ 제1종 보통면허 및\
                    \ 소형견인차면허\",\"④ 제2종 보통면허 및 대형견인차면허\"],\"answers\":[4],\"commentary\"\
                    :\"도로교통법 시행규칙 별표18 총중량 750킬로그램을 초과하는 3톤 이하의 피견인 자동차를 견인하기 위해서는\
                    \ 견인 하는 자동차를 운전할 수 있는 면허와 소형견인차면허 또는 대형견인차면허를 가지고 있어야 한다.\",\"\
                    commentary_images_in\":[{\"url\":\"https://examlab-image.s3.ap-northeast-2.amazonaws.com/test_images/image3.png\"\
                    ,\"description\":\"설명3\",\"attribute\":\"속성3\"}],\"commentary_images_out\"\
                    :[{\"url\":\"https://examlab-image.s3.ap-northeast-2.amazonaws.com/test_images/image4.png\"\
                    ,\"description\":\"설명4\",\"attribute\":\"속성4\"}],\"tags\":{}}}"
components:
  schemas:
...
    api-v1-questions-420259584:
      required:
      - answers
      - commentary
      - id
      - options
      - question
      - tags
      type: object
      properties:
        question:
          type: string
          description: 변경될 질문 내용
        answers:
          type: array
          description: 변경될 정답
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        options:
          type: array
          description: 변경될 보기 내용
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        id:
          type: string
          description: 문제 ID
        commentary:
          type: string
          description: 변경될 해설
        tags:
          type: object
          properties:
            '*':
              type: array
              description: 변경될 문제 태그 value
              nullable: true
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
          description: 변경될 문제 태그
...
    api-v1-exams-examId-questions-1876066633:
      required:
      - questions
      - size
      type: object
      properties:
        size:
          type: number
          description: 문제의 개수
        questions:
          type: array
          description: 문제 목록
          items:
            required:
            - answers
            - commentary
            - commentary_images_in
            - commentary_images_out
            - id
            - options
            - question
            - question_images_in
            - question_images_out
            - tags
            - type
            type: object
            properties:
              question:
                type: string
                description: 문제 내용
              question_images_out:
                type: array
                description: 문제 정답 이미지 목록
                items:
                  type: object
                  properties:
                    description:
                      type: string
                      description: 이미지 설명
                      nullable: true
                    attribute:
                      type: string
                      description: 이미지 속성
                      nullable: true
                    url:
                      type: string
                      description: 이미지 URL
                      nullable: true
              question_images_in:
                type: array
                description: 문제 설명 이미지 목록
                items:
                  type: object
                  properties:
                    description:
                      type: string
                      description: 이미지 설명
                      nullable: true
                    attribute:
                      type: string
                      description: 이미지 속성
                      nullable: true
                    url:
                      type: string
                      description: 이미지 URL
                      nullable: true
              answers:
                type: array
                description: 정답 목록
                items:
                  oneOf:
                  - type: object
                  - type: boolean
                  - type: string
                  - type: number
              options:
                type: array
                description: 보기 목록
                items:
                  oneOf:
                  - type: object
                  - type: boolean
                  - type: string
                  - type: number
              id:
                type: string
                description: 문제의 고유 식별자
              type:
                type: string
                description: 문제의 유형
              commentary_images_in:
                type: array
                description: 해설 설명 이미지 목록
                items:
                  type: object
                  properties:
                    description:
                      type: string
                      description: 이미지 설명
                      nullable: true
                    attribute:
                      type: string
                      description: 이미지 속성
                      nullable: true
                    url:
                      type: string
                      description: 이미지 URL
                      nullable: true
              commentary_images_out:
                type: array
                description: 해설 이미지 목록
                items:
                  type: object
                  properties:
                    description:
                      type: string
                      description: 이미지 설명
                      nullable: true
                    attribute:
                      type: string
                      description: 이미지 속성
                      nullable: true
                    url:
                      type: string
                      description: 이미지 URL
                      nullable: true
              commentary:
                type: string
                description: 문제 해설
              tags:
                type: object
                properties:
                  '*':
                    type: array
                    description: 문제의 카테고리 태그 정보
                    nullable: true
                    items:
                      oneOf:
                      - type: object
                      - type: boolean
                      - type: string
                      - type: number
                description: 문제의 태그 맵 정보
...
    api-v1-questions486549215:
      type: object
...